+++ CHANGELOG.md
@@ -1,5 +1,15 @@
+## 0.0.3
+
+* Add `SerializationException` => `SerializationFailure`
+* Add `NetworkConnetionException` => `NetworkConnetionFailure`
+* Add `WrongPasswordException ` => `WrongPasswordFailure`
+* Add `EmailAlreadyExistException ` => `EmailAlreadyExistFailure`
+* Add utils `exceptionToFailure` 
+
+## 0.0.2
+
+* Refactoring & export exceptions
+
 ## 0.0.1
 
 * Add networkException, ServerException,UnauthorizedException,UnknownException
-## 0.0.2 
-* Refactoring & export exceptions